address git pull concerns - cleaner code and new config file
