+++ ui/src/pages/Page/Page.scss
@@ -1,6 +1,7 @@
 @import "../../styles/Constants";
 
 
+
 body {
   background-color: $lightBlueBackground;
